+++ src/main/java/kitchenpos/menu/exception/InvalidMenuGroupException.java
@@ -1,4 +1,6 @@
-package kitchenpos.common.exception;
+package kitchenpos.menu.exception;
+
+import kitchenpos.common.exception.BadRequestException;
 
 public class InvalidMenuGroupException extends BadRequestException {
