+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -3,10 +3,8 @@ package kitchenpos.order.application;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.common.exception.InvalidMenuException;
-import kitchenpos.common.exception.InvalidOrderException;
-import kitchenpos.common.exception.InvalidTableException;
 import kitchenpos.menu.dao.MenuDao;
+import kitchenpos.menu.exception.InvalidMenuException;
 import kitchenpos.order.application.dto.request.OrderLineItemRequest;
 import kitchenpos.order.application.dto.request.OrderRequest;
 import kitchenpos.order.application.dto.request.OrderStatusRequest;
@@ -17,8 +15,10 @@ import kitchenpos.order.dao.OrderLineItemDao;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.order.exception.InvalidOrderException;
 import kitchenpos.table.dao.OrderTableDao;
 import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.exception.InvalidTableException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.CollectionUtils;
