+++ src/test/java/kitchenpos/domain/TableGroupTest.java
@@ -4,9 +4,9 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import kitchenpos.common.exception.InvalidTableException;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.TableGroup;
+import kitchenpos.table.exception.InvalidTableException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
